GH #1569 - Bugfix: EcephysProjectCache failing on write
When the 'EcephysProjectCache' is constructed directly rather
than using the 'classmethod' constructors (e.g. 'from_lims',
'from_warehouse'), the default for the 'stream_writer' param  is
incompatible with the default for the 'RmaEngine' in the 'fetch_api'.
This update adds the appropriate stream writer as a
static method in 'HttpEngine' and 'AsyncHttpEngine', which is used
as the default if 'stream_writer' is not passed in the
class constructor for 'EcephysProjectCache'. This resolves
the cache instance trying to use an incompatible stream writer
when constructed using default values in the '__init__' method.
It is still possible to pass a custom callable to 'stream_writer'
when constructing 'EcephysProjectCache' if desired.

This update also adds additional notes to the user in the
docstring for 'EcephysProjectCache', recommending using the
classmethod constructors ('from_lims', 'from_warehouse', 'fixed')
to create the 'EcephysProjectCache' rather than constructing it
directly.